+++ Code/Legacy/CryCommon/CryString.h
@@ -1266,7 +1266,7 @@ inline void CryStringT<T>::resize(size_type nCount, value_type _Ch)
     }
     else if (nCount < length())
     {
-        _header()->nLength = nCount;
+        _header()->nLength = static_cast<int>(nCount);
         m_str[length()] = 0; // Make null terminated string.
     }
 }
@@ -1971,7 +1971,7 @@ inline CryStringT<T>& CryStringT<T>::erase(size_type nIndex, size_type nCount)
         _MakeUnique();
         size_type nNumToCopy = length() - (nIndex + nCount) + 1;
         _move(m_str + nIndex, m_str + nIndex + nCount, nNumToCopy);
-        _header()->nLength = length() - nCount;
+        _header()->nLength = static_cast<int>(length() - nCount);
     }
 
     return *this;
@@ -2013,7 +2013,7 @@ inline CryStringT<T>& CryStringT<T>::insert(size_type nIndex, size_type nCount, 
 
     _move(m_str + nIndex + nCount, m_str + nIndex, (nNewLength - nIndex - nCount) + 1);
     _set(m_str + nIndex, ch, nCount);
-    _header()->nLength = nNewLength;
+    _header()->nLength = static_cast<int>(nNewLength);
     CRY_STRING_DEBUG(m_str)
 
     return *this;
@@ -2050,7 +2050,7 @@ inline CryStringT<T>& CryStringT<T>::insert(size_type nIndex, const_str pstr, si
 
         _move(m_str + nIndex + nInsertLength, m_str + nIndex, (nNewLength - nIndex - nInsertLength + 1));
         _copy(m_str + nIndex, pstr, nInsertLength);
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
         m_str[length()] = 0;
     }
     CRY_STRING_DEBUG(m_str)
@@ -2164,7 +2164,7 @@ inline CryStringT<T>& CryStringT<T>::replace(const_str strOld, const_str strNew)
             }
             strStart += _strlen(strStart) + 1;
         }
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
     }
     CRY_STRING_DEBUG(m_str)
 
@@ -2289,7 +2289,7 @@ inline CryStringT<T>& CryStringT<T>::TrimRight(const value_type* sCharSet)
         // Just shrink length of the string.
         size_type nNewLength = (size_type)(str - m_str) + 1; // m_str can change in _MakeUnique
         _MakeUnique();
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
         m_str[nNewLength] = 0;
     }
 
@@ -2317,7 +2317,7 @@ inline CryStringT<T>& CryStringT<T>::TrimRight()
         // Just shrink length of the string.
         size_type nNewLength = (size_type)(str - m_str) + 1; // m_str can change in _MakeUnique
         _MakeUnique();
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
         m_str[nNewLength] = 0;
     }
 
@@ -2353,7 +2353,7 @@ inline CryStringT<T>& CryStringT<T>::TrimLeft(const value_type* sCharSet)
         _MakeUnique();
         size_type nNewLength = length() - nOff;
         _move(m_str, m_str + nOff, nNewLength + 1);
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
         m_str[nNewLength] = 0;
     }
 
@@ -2376,7 +2376,7 @@ inline CryStringT<T>& CryStringT<T>::TrimLeft()
         _MakeUnique();
         size_type nNewLength = length() - nOff;
         _move(m_str, m_str + nOff, nNewLength + 1);
-        _header()->nLength = nNewLength;
+        _header()->nLength = static_cast<int>(nNewLength);
         m_str[nNewLength] = 0;
     }
